split user migration tests
