+++ script/ci/ordered_tests.txt
@@ -4,9 +4,9 @@ spec/models/user_part_database_and_cache_spec.rb
 spec/models/user_part_organization_spec.rb
 spec/models/user_part_plans_and_services_spec.rb
 spec/models/user_part_validation_and_authentication_spec.rb
-spec/models/carto/user_migration_spec_part_1.rb
-spec/models/carto/user_migration_spec_part_2.rb
-spec/models/carto/user_migration_spec_part_3.rb
+spec/models/carto/user_migration_api_key_spec.rb
+spec/models/carto/user_migration_rollback_spec.rb
+spec/models/carto/user_migration_base_spec.rb
 spec/requests/carto/api/visualizations_controller_spec_part_1.rb
 spec/requests/carto/api/visualizations_controller_spec_part_2.rb
 spec/requests/carto/api/visualizations_controller_spec_part_3.rb
